Switched to category-id and refactored package name
