+++ config-xml/schema/windup-jboss-ruleset.xsd
@@ -282,7 +282,7 @@
                 <xs:element ref="quickfix" minOccurs="0" maxOccurs="unbounded"/>
             </xs:sequence>
             <xs:attribute name="title" type="xs:string" use="required"/>
-            <xs:attribute name="categoryID" type="category-id-type" use="optional" />
+            <xs:attribute name="category-id" type="category-id-type" use="optional" />
             <xs:attribute name="in" type="xs:string" use="optional" />
             <xs:attribute name="effort" type="xs:byte" use="optional" />
             <!-- <xs:attribute type="xs:string" name="message" use="optional"/> -->
@@ -319,7 +319,7 @@
             </xs:sequence>
             <xs:attribute name="title" type="xs:string" use="required"/>
             <xs:attribute name="effort" type="xs:byte" use="optional" />
-            <xs:attribute name="categoryID" type="category-id-type" use="optional" />
+            <xs:attribute name="category-id" type="category-id-type" use="optional" />
             <xs:attribute name="of" type="xs:string" use="optional" />
         </xs:complexType>
     </xs:element>
