+++ reporting/api/src/main/java/org/jboss/windup/reporting/category/IssueCategory.java
@@ -1,4 +1,4 @@
-package org.jboss.windup.reporting.severity;
+package org.jboss.windup.reporting.category;
 
 /**
  * <p>
