+++ reporting/api/src/main/java/org/jboss/windup/reporting/category/LoadIssueCategoriesRuleProvider.java
@@ -1,4 +1,4 @@
-package org.jboss.windup.reporting.severity;
+package org.jboss.windup.reporting.category;
 
 import java.io.IOException;
 import java.nio.file.FileVisitResult;
