+++ reporting/api/src/main/java/org/jboss/windup/reporting/config/HintText.java
@@ -1,7 +1,7 @@
 package org.jboss.windup.reporting.config;
 
 import java.util.Set;
-import org.jboss.windup.reporting.severity.IssueCategory;
+import org.jboss.windup.reporting.category.IssueCategory;
 import org.ocpsoft.rewrite.config.OperationBuilder;
 
 /**
