+++ reporting/impl/src/main/resources/reports/templates/application_list.ftl
@@ -19,7 +19,7 @@
     <#-- appReport : ApplicationReportModel -->
 
     <#assign allTraversal  = getProjectTraversal(appReport.projectModel, 'all')>
-    <#assign incidentCountBySeverity = getEffortCountForProjectBySeverity(event, allTraversal, true)>
+    <#assign incidentCountByCategory = getEffortCountForProjectByIssueCategory(event, allTraversal, true)>
 
     <#include "include/effort_util.ftl">
     <#assign allTraversal  = getProjectTraversal(appReport.projectModel, 'all')>
@@ -61,10 +61,10 @@
                         <td colspan="2">Number of incidents</td>
                     </tr>
                     <#assign totalIncidents = 0 >
-                    <#list incidentCountBySeverity?keys as issueCategory>
-                        <#assign totalIncidents = totalIncidents + incidentCountBySeverity?api.get(issueCategory) >
+                    <#list incidentCountByCategory?keys as issueCategory>
+                        <#assign totalIncidents = totalIncidents + incidentCountByCategory?api.get(issueCategory) >
                         <tr>
-                            <td class="count">${incidentCountBySeverity?api.get(issueCategory)}</td>
+                            <td class="count">${incidentCountByCategory?api.get(issueCategory)}</td>
                             <td class="label_">${issueCategory.name}</td>
                         </tr>
                     </#list>
