+++ reporting/tests/src/test/java/org/jboss/windup/reporting/LoadIssueCategoriesRuleProviderTest.java
@@ -14,8 +14,8 @@ import org.jboss.windup.config.RuleSubset;
 import org.jboss.windup.config.loader.RuleLoaderContext;
 import org.jboss.windup.graph.GraphContext;
 import org.jboss.windup.graph.GraphContextFactory;
-import org.jboss.windup.reporting.severity.IssueCategoryRegistry;
-import org.jboss.windup.reporting.severity.LoadIssueCategoriesRuleProvider;
+import org.jboss.windup.reporting.category.IssueCategoryRegistry;
+import org.jboss.windup.reporting.category.LoadIssueCategoriesRuleProvider;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
